[#127][be][user]refactor: Auth 패키지 User로 이동 및 수정
