--- back/src/main/java/com/rouby/auth/application/facade/AuthFacade.java
@@ -1,45 +1,0 @@
-package com.rouby.auth.application.facade;
-
-import com.rouby.auth.application.dto.request.LoginCommand;
-import com.rouby.auth.application.dto.response.LoginInfo;
-import com.rouby.auth.jwt.JwtTokenProvider;
-import com.rouby.common.exception.CustomException;
-import com.rouby.user.application.service.UserReadService;
-import com.rouby.user.application.exception.UserErrorCode;
-import com.rouby.user.domain.entity.User;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-/**
- * @Date : 2025. 07. 08.
- *
- * @author : hanjihoon
- */
-@Service
-@RequiredArgsConstructor
-public class AuthFacade {
-
-  private final UserReadService userReadService;
-  private final PasswordEncoder passwordEncoder;
-  private final JwtTokenProvider jwtTokenProvider;
-
-  @Transactional(readOnly = true)
-  public LoginInfo login(LoginCommand command){
-
-    User user = userReadService.findByEmail(command.email());
-
-    if (!passwordEncoder.matches(command.password(), user.getPassword())) {
-      throw CustomException.from(UserErrorCode.INVALID_USER_PASSWORD);
-    }
-
-    return new LoginInfo(jwtTokenProvider.createAccessToken(
-        user.getId().toString(),
-        user.getRole().toString(),
-        user.getEmail()));
-  }
-
-
-
-}
