+++ back/src/main/java/com/rouby/user/application/dto/info/LoginInfo.java
@@ -1,4 +1,4 @@
-package com.rouby.auth.application.dto.response;
+package com.rouby.user.application.dto.info;
 
 /**
  * @Date : 2025. 07. 08.
