+++ back/src/main/java/com/rouby/user/application/service/UserReadService.java
@@ -1,10 +1,13 @@
 package com.rouby.user.application.service;
 
 import static com.rouby.user.application.exception.UserErrorCode.INVALID_USER;
+import static com.rouby.user.application.exception.UserErrorCode.INVALID_USER_PASSWORD;
 
-import com.rouby.common.exception.CustomException;
+import com.rouby.user.application.dto.command.LoginCommand;
+import com.rouby.user.application.exception.UserException;
 import com.rouby.user.domain.entity.User;
 import com.rouby.user.domain.repository.UserRepository;
+import com.rouby.user.domain.service.UserPasswordEncoder;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -14,11 +17,25 @@ import org.springframework.transaction.annotation.Transactional;
 public class UserReadService {
 
   private final UserRepository userRepository;
+  private final UserPasswordEncoder passwordEncoder;
+
+
+  @Transactional(readOnly = true)
+  public User validUser(LoginCommand command) {
+    User user = findByEmail(command.email());
+
+    if (!passwordEncoder.matches(command.password(), user.getPassword())) {
+      throw UserException.from(INVALID_USER_PASSWORD);
+    }
+
+    return user;
+  }
+
 
   @Transactional(readOnly = true)
   public User findByEmail(String email) {
     return userRepository.findByEmail(email)
-        .orElseThrow(() -> CustomException.from(INVALID_USER));
+        .orElseThrow(() -> UserException.from(INVALID_USER));
   }
 
   public boolean alreadyExistsEmail(String email) {
