+++ back/src/main/java/com/rouby/user/infrastructure/security/handler/JwtAuthenticationEntryPoint.java
@@ -1,4 +1,4 @@
-package com.rouby.auth.handler;
+package com.rouby.user.infrastructure.security.handler;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
