+++ back/src/test/http/user.http
@@ -15,3 +15,12 @@ content-type: application/json
   "email": "invalid@hidle.com",
   "password": "qwe3"
 }
+
+### 로그인
+POST {{url}}/api/v1/auth/login
+content-type: application/json
+
+{
+  "email": "valid@hidle.com",
+  "password": "aaaaaa11!"
+}
