+++ back/src/test/java/com/rouby/common/support/ControllerTestSupport.java
@@ -5,9 +5,8 @@ import com.rouby.common.config.WebConfig;
 import com.rouby.common.config.WebMvcConfig;
 import com.rouby.common.exception.GlobalExceptionHandler;
 import com.rouby.common.resolver.CustomPageableArgumentResolver;
-import com.rouby.auth.application.facade.AuthFacade;
-import com.rouby.auth.filter.JwtAuthenticationFilter;
-import com.rouby.auth.presentation.AuthController;
+import com.rouby.user.infrastructure.security.filter.JwtAuthenticationFilter;
+import com.rouby.user.presentation.AuthController;
 import com.rouby.schedule.application.facade.ScheduleFacade;
 import com.rouby.schedule.presentation.ScheduleController;
 import org.springframework.beans.factory.annotation.Autowired;
